requirements.txt
appdirs==1.4.4
blessed==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
idna==3.6
importlib_resources==6.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sentry-sdk==1.44.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
torch==2.2.0+cu118
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.5
wcwidth==0.2.13
zipp==3.18.1
